PAY-14

+++ app/Http/Controllers/Api/ContactController.php
@@ -152,6 +152,9 @@ class ContactController extends AbstractRestAPIController
      */
     public function indexMyContact(IndexRequest $request)
     {
+        if (!Gate::allows('permission', config('api.contact.index'))) {
+            return $this->sendJsonResponse(false, 'You need to upgrade platform package', [], 403);
+        }
         try {
             $filters = $request->filter;
             if (!empty($filters['uuids_in']) && !empty($filters['uuids_not_in'])) {
@@ -189,6 +192,9 @@ class ContactController extends AbstractRestAPIController
      */
     public function storeMyContact(MyContactRequest $request)
     {
+        if (!Gate::allows('permission', config('api.contact.create'))) {
+            return $this->sendJsonResponse(false, 'You need to upgrade platform package', [], 403);
+        }
         $model = $this->service->create(array_merge($request->all(), [
             'user_uuid' => auth()->user()->getkey(),
         ]));
@@ -208,6 +214,9 @@ class ContactController extends AbstractRestAPIController
      */
     public function showMyContact($id)
     {
+        if (!Gate::allows('permission', config('api.contact.show'))) {
+            return $this->sendJsonResponse(false, 'You need to upgrade platform package', [], 403);
+        }
         $model = $this->myService->findMyContactByKeyOrAbort($id);
 
         return $this->sendOkJsonResponse(
@@ -222,6 +231,9 @@ class ContactController extends AbstractRestAPIController
      */
     public function editMyContact(UpdateMyContactRequest $request, $id)
     {
+        if (!Gate::allows('permission', config('api.contact.edit'))) {
+            return $this->sendJsonResponse(false, 'You need to upgrade platform package', [], 403);
+        }
         $model = $this->myService->findMyContactByKeyOrAbort($id);
 
         $this->service->update($model, array_merge($request->except("points"), [
@@ -243,6 +255,9 @@ class ContactController extends AbstractRestAPIController
      */
     public function destroyMyContact($id)
     {
+        if (!Gate::allows('permission', config('api.contact.delete'))) {
+            return $this->sendJsonResponse(false, 'You need to upgrade platform package', [], 403);
+        }
         $this->myService->deleteMyContactByKey($id);
 
         return $this->sendOkJsonResponse();
